# contracts/openapi/talli-v1.json
{
  "components": {
    "schemas": {
      "CompanyContext": {
        "properties": {
          "aal": {
            "enum": [
              "aal1",
              "aal2"
            ],
            "title": "Aal",
            "type": "string"
          },
          "address": {
            "title": "Address",
            "type": "string"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "title": "Createdby",
            "type": "string"
          },
          "entityType": {
            "title": "Entitytype",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "identityConfirmedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identityconfirmedat"
          },
          "identityLockedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identitylockedat"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "orgNumber": {
            "title": "Orgnumber",
            "type": "string"
          },
          "postalCode": {
            "title": "Postalcode",
            "type": "string"
          },
          "resourceScope": {
            "enum": [
              "workspace",
              "owner",
              "owner_sensitive"
            ],
            "title": "Resourcescope",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "reviewer",
              "read_only"
            ],
            "title": "Role",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "statusText": {
            "title": "Statustext",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orgNumber",
          "name",
          "entityType",
          "address",
          "postalCode",
          "city",
          "statusText",
          "source",
          "createdBy",
          "identityConfirmedAt",
          "identityLockedAt",
          "createdAt",
          "role",
          "resourceScope",
          "aal"
        ],
        "title": "CompanyContext",
        "type": "object"
      },
      "CompanyContextResponse": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyContext"
            },
            "title": "Companies",
            "type": "array"
          },
          "selectedCompany": {
            "$ref": "#/components/schemas/CompanyContext"
          }
        },
        "required": [
          "selectedCompany",
          "companies"
        ],
        "title": "CompanyContextResponse",
        "type": "object"
      },
      "SystemBoundaryStatus": {
        "properties": {
          "apiVersion": {
            "title": "Apiversion",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "AVAILABLE",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "apiVersion",
          "service",
          "status"
        ],
        "title": "SystemBoundaryStatus",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Talli web-to-backend production boundary",
    "title": "Talli API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/company-access/context": {
      "get": {
        "operationId": "companyAccessGetSelectedContext",
        "parameters": [
          {
            "in": "query",
            "name": "company_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "description": "Optional caller-provided correlation identifier.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyContextResponse"
                }
              }
            },
            "description": "Selected company context.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Company context request failed.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Company context request failed.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Company context request failed.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Company context request failed.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Company context request failed.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Selected Company Context",
        "tags": [
          "company-access"
        ]
      }
    },
    "/api/v1/system-boundary/tracer": {
      "get": {
        "operationId": "systemBoundaryGetTracerStatus",
        "parameters": [
          {
            "description": "Optional caller-provided correlation identifier.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemBoundaryStatus"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "An unexpected backend failure occurred.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "The backend boundary is temporarily unavailable.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier for the request and response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get System Boundary Status",
        "tags": [
          "system-boundary"
        ]
      }
    }
  }
}